# perf(core): use raw v8 get/set data