Instead of rusty_v8's Annex to reduce lookup overhead

```
Before:
test bench_is_proxy   ... bench:      20,051 ns/iter (+/- 860)
test bench_op_async   ... bench:     391,111 ns/iter (+/- 5,354)
test bench_op_nop     ... bench:      52,359 ns/iter (+/- 1,942)
test bench_op_pi_json ... bench:      54,409 ns/iter (+/- 1,326)

After:
test bench_is_proxy   ... bench:      20,292 ns/iter (+/- 480)
test bench_op_async   ... bench:     385,609 ns/iter (+/- 5,775)
test bench_op_nop     ... bench:      46,042 ns/iter (+/- 1,785)
test bench_op_pi_json ... bench:      49,903 ns/iter (+/- 668)
```

Blocked on https://github.com/denoland/rusty_v8/pull/911