@aapoalas 2022-03-12
Is the Annex actually even needed after this? It seems to only contain the `create_param_allocation` (snapshot params?) which is never accessed from it, only kept by it. If that memory is needed to be kept, it looks like it could be just another internal slot instead of indirected through the Annex.